requirements.txt
APScheduler==3.9.1
asgiref==3.4.1
astroid==2.8.0
attrs==21.4.0
autobahn==22.6.1
Automat==20.2.0
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.4
charset-normalizer==2.1.0
constantly==15.1.0
coverage==6.4.2
cryptography==37.0.4
daphne==3.0.2
Django==3.2.12
django-filter==21.1
django-pandas==0.6.6
django-revproxy==0.10.0
django-tables2==2.4.1
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
influxdb==5.3.1
influxdb-client==1.30.0
isort==5.9.3
jplephem==2.17
lazy-object-proxy==1.6.0
mccabe==0.6.1
msgpack==1.0.4
numpy==1.23.1
pandas==1.3.5
platformdirs==2.4.0
psycopg2==2.8.6
py4j==0.10.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pylint==2.11.1
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
Rx==3.2.0
service-identity==21.1.0
sgp4==2.21
six==1.16.0
skyfield==1.41
sqlparse==0.4.2
toml==0.10.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==3.10.0.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
uWSGI==2.0.20
wrapt==1.12.1
zope.interface==5.4.0
